Backend/requirements.txt
pillow
django
pandas
Prophet
requests
psycopg2
python-dotenv
beautifulsoup4
djangorestframework
django-simple-captcha
djangorestframework-simplejwt
